Change class variables to instance variables
TempestConf has a set of class variables that do not need to be
class variables because there are no multiple instances of
TempestConf class that could share the value.

The reason why we need to move the class variables to instance
variables is that some unit tests under certain conditions were
failing because multiple unit tests were changing the
TempestConf.priority_sectionkeys value.

Change-Id: Ic6f51be112aa9f93904fcab289d78a7e06f20f62