Fix opendev-tox-docs job
With the upgrade of tox to a newer version we started to encounter
failures in the opendev-tox-docs job. It looks like the interpretation
of the variable skipdist changed.

A similar problem was solved in tempest repository by removing the
skipsdist option from tox [1]. It looks like a similar solution
works for python-tempestconf as well.

[1] https://review.opendev.org/c/openstack/tempest/+/868527

Depends-On: Ie5ea7f9305c7e7611f205d2e0703849bd6321a68
Change-Id: I9ee6570b72c10ea0466819008135ddebe891a445