Handle when there is no volume backend available
We query the volume pools to check if the user has permissions to list
the backend storage, but we do not handle when the volume backend is not
available, only when the user is not allowed. This patch adds the
ServiceError exception to handle it and set the backend properly.

Change-Id: I0fd61813dbadf5ef3a0db55791e3b261d4c9d885