Unblocking CI by n-v a few jobs
The commits marks several jobs as n-v to unblock the CI. The jobs
already have fixes proposed and are in the process of review.
Right now we wait on a change in Tempest which may take serveral
days to get merged.
The patch also fixes an issue with pep8 - too long line - seems
like the sensitivity of pep8 was increased because previously
the line was ok with pep8.

Change-Id: Icea222faadfe22ab029d0321fd68c3e6f1bbc0b4